refactor(platform): move steam and cdrom into platform facades, completing platform seam
