+++ src/kfx/platform/IPlatform.h
@@ -1,6 +1,9 @@
 #ifndef IPLATFORM_H
 #define IPLATFORM_H
 
+#include "bflib_basics.h" // TbBool
+#include "bflib_sound.h"  // SoundVolume
+
 class IWindowSystem;
 struct TbFileFind;
 struct TbFileEntry;
@@ -19,6 +22,17 @@ public:
     virtual const char* GetWineHost() const = 0;    // nullptr when not running under Wine
 
     virtual TbFileFind* FileFindFirst(const char* filespec, TbFileEntry* entry) = 0;
+
+    // ----- Redbook (CD) audio -----
+    virtual void   SetRedbookVolume(SoundVolume vol) = 0;
+    virtual TbBool PlayRedbookTrack(int track) = 0;
+    virtual void   PauseRedbookTrack() = 0;
+    virtual void   ResumeRedbookTrack() = 0;
+    virtual void   StopRedbookTrack() = 0;
+
+    // ----- Steam ----- (Linux stubbed; real libsteam_api.so is future work)
+    virtual int  InitSteam() = 0;   // 0 ok, -1 unsupported, >0 failure
+    virtual void ShutdownSteam() = 0;
 
     /** Initialise the display subsystem. Per-OS: Windows adjusts SDL hints
      *  before SDL_Init, Linux initialises plainly. Returns false on failure. */
