+++ src/kfx/platform/PlatformManager.cpp
@@ -15,6 +15,8 @@
 #include "kfx/platform/FileFind.h"
 #include "platform.h"
 #include "bflib_fileio.h"
+#include "cdrom.h"
+#include "steam_api.hpp"
 #include "post_inc.h"
 
 /******************************************************************************/
@@ -33,10 +35,10 @@ IPlatform* GetPlatform()
 
 /******************************************************************************/
 
-extern "C" const char * get_os_version(void)   { return GetPlatform()->GetOSVersion(); }
-extern "C" const void * get_image_base(void)    { return GetPlatform()->GetImageBase(); }
-extern "C" const char * get_wine_version(void)  { return GetPlatform()->GetWineVersion(); }
-extern "C" const char * get_wine_host(void)     { return GetPlatform()->GetWineHost(); }
+extern "C" const char * PlatformManager_GetOSVersion(void)   { return GetPlatform()->GetOSVersion(); }
+extern "C" const void * PlatformManager_GetImageBase(void)   { return GetPlatform()->GetImageBase(); }
+extern "C" const char * PlatformManager_GetWineVersion(void) { return GetPlatform()->GetWineVersion(); }
+extern "C" const char * PlatformManager_GetWineHost(void)    { return GetPlatform()->GetWineHost(); }
 
 /******************************************************************************/
 
@@ -63,6 +65,17 @@ extern "C" void LbFileFindEnd(struct TbFileFind * ffind)
 {
     delete ffind;
 }
+
+/******************************************************************************/
+
+extern "C" void   SetRedbookVolume(SoundVolume value) { GetPlatform()->SetRedbookVolume(value); }
+extern "C" TbBool PlayRedbookTrack(int track)         { return GetPlatform()->PlayRedbookTrack(track); }
+extern "C" void   PauseRedbookTrack(void)             { GetPlatform()->PauseRedbookTrack(); }
+extern "C" void   ResumeRedbookTrack(void)            { GetPlatform()->ResumeRedbookTrack(); }
+extern "C" void   StopRedbookTrack(void)              { GetPlatform()->StopRedbookTrack(); }
+
+extern "C" int  steam_api_init(void)     { return GetPlatform()->InitSteam(); }
+extern "C" void steam_api_shutdown(void) { GetPlatform()->ShutdownSteam(); }
 
 /******************************************************************************/
